Delete repository due to fatal error: object head
I had a fatal error that was not allowing me to commit changes to the
repo. The quickest way to solve this was to delete the repo and make a
new one. So far I had finished the basic structure for the home page and
the styling for the header and hero. I will try to recreate this from
memory, this will also be an opportunity to make any improvements to the
styling and structure as well. The only file that was saved was the
index file, my styling is completely gone ¯\_(ツ)_/¯